tweaks

+++ examples/dqn_orderbook/symbol_tuner.py
@@ -293,11 +293,12 @@ class SymbolTuner(StudyWrapper):
         self.trial = trial
 
         try:
-            exported_model_path = self.study.best_trial.user_attrs['exported_model_path']
-            if self.export_best and self.study.best_trial.value > self.min_capital:
-                self.save_best_params()
-                shutil.rmtree(self.best_model_dir, ignore_errors=True)
-                shutil.copytree(exported_model_path, self.best_model_dir)
+            if 'exported_model_path' in self.study.best_trial.user_attrs:
+                exported_model_path = self.study.best_trial.user_attrs['exported_model_path']
+                if self.export_best and self.study.best_trial.value > self.min_capital:
+                    self.save_best_params()
+                    shutil.rmtree(self.best_model_dir, ignore_errors=True)
+                    shutil.copytree(exported_model_path, self.best_model_dir)
         except ValueError as err:
             pass
